move wrapper to factory
