--- Scripts/ThirdPartyServices/AnalyticEvents/Falcon/FalconAnalyticWrapper.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: f8b89fbbeacdd914a9d55f4c1aeffb5b
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
